+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `customCssClasses` to form element validation
+- `requireMessage` to form element validation
 
 ## [0.19.9] - 2022-07-29